Remove handle-request-for as unnecessary injectable
